Start figuring out trpc
